+++ src/features/browse/Browse.types.ts
@@ -10,8 +10,7 @@ import { SourceIdInfo } from '@/features/source/Source.types.ts';
 
 export type MetadataBrowseSettings = {
     hideLibraryEntries: boolean;
-    extensionLanguages: string[];
-    sourceLanguages: string[];
+    browseLanguages: string[];
     showNsfw: boolean;
     lastUsedSourceId: SourceIdInfo['id'] | null;
     shouldShowOnlySourcesWithResults: boolean;