+++ src/features/settings/Settings.constants.ts
@@ -68,8 +68,7 @@ export const SERVER_SETTINGS_METADATA_DEFAULT: MetadataServerSettings = {
 
     // browse
     hideLibraryEntries: false,
-    extensionLanguages: getDefaultLanguages(),
-    sourceLanguages: getDefaultLanguages(),
+    browseLanguages: getDefaultLanguages(),
     showNsfw: true,
     lastUsedSourceId: null,
     shouldShowOnlySourcesWithResults: true,